move server to library
