--- psiresp/tests/server.py
@@ -1,7 +1,0 @@
-import pytest
-
-import qcfractal.interface as ptl
-from psiresp.testing import TemporaryPostgres
-from qcfractal import FractalSnowflake
-
-from psiresp.tests.datafiles import POSTGRES_SERVER_BACKUP
